Benchmarking postgres reads (#109)
