+++ flow/connectors/postgres/qrep.go
@@ -1,17 +1,12 @@
 package connpostgres
 
 import (
-	"database/sql"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/PeerDB-io/peer-flow/generated/protos"
 	"github.com/PeerDB-io/peer-flow/model"
 	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgconn"
-	"github.com/jackc/pgx/v5/pgtype"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -93,111 +88,9 @@ func (c *PostgresConnector) GetQRepPartitions(config *protos.QRepConfig,
 	return partitions, nil
 }
 
-func mapRowToQRecord(row pgx.Row, fds []pgconn.FieldDescription) (*model.QRecord, error) {
-	record := &model.QRecord{}
-
-	scanArgs := make([]interface{}, len(fds))
-	for i := range scanArgs {
-		if fds[i].DataTypeOID == pgtype.BoolOID {
-			scanArgs[i] = new(sql.NullBool)
-		} else if fds[i].DataTypeOID == pgtype.TimestampOID || fds[i].DataTypeOID == pgtype.TimestamptzOID {
-			scanArgs[i] = new(sql.NullTime)
-		} else {
-			scanArgs[i] = new(sql.RawBytes)
-		}
-	}
-
-	err := row.Scan(scanArgs...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to scan row: %w", err)
-	}
-
-	for i, fd := range fds {
-		oid := fd.DataTypeOID
-		var val model.QValue
-		var err error
-
-		switch oid {
-		case pgtype.TimestampOID, pgtype.TimestamptzOID:
-			nullTime := scanArgs[i].(*sql.NullTime)
-			if nullTime.Valid {
-				et, err := model.NewExtendedTime(nullTime.Time, model.DateTimeKindType, "")
-				if err != nil {
-					return nil, fmt.Errorf("failed to create extended time: %w", err)
-				}
-				val = model.QValue{Kind: model.QValueKindETime, Value: et}
-			} else {
-				val = model.QValue{Kind: model.QValueKindETime, Value: nil}
-			}
-		case pgtype.BoolOID:
-			nullBool := scanArgs[i].(*sql.NullBool)
-			if nullBool.Valid {
-				val = model.QValue{Kind: model.QValueKindBoolean, Value: nullBool.Bool}
-			} else {
-				val = model.QValue{Kind: model.QValueKindBoolean, Value: nil}
-			}
-		default:
-			rawBytes := scanArgs[i].(*sql.RawBytes)
-			val, err = parseField(oid, rawBytes)
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse field: %w", err)
-			}
-		}
-
-		(*record)[fd.Name] = val
-	}
-
-	return record, nil
-}
-
-func parseField(oid uint32, rawBytes *sql.RawBytes) (model.QValue, error) {
-	switch oid {
-	case pgtype.Int4OID, pgtype.Int8OID:
-		val, err := strconv.ParseInt(string(*rawBytes), 10, 64)
-		if err != nil {
-			return model.QValue{}, err
-		}
-		return model.QValue{Kind: model.QValueKindInteger, Value: val}, nil
-	case pgtype.Float4OID, pgtype.Float8OID:
-		val, err := strconv.ParseFloat(string(*rawBytes), 64)
-		if err != nil {
-			return model.QValue{}, err
-		}
-		return model.QValue{Kind: model.QValueKindFloat, Value: val}, nil
-	case pgtype.BoolOID:
-		val, err := strconv.ParseBool(string(*rawBytes))
-		if err != nil {
-			return model.QValue{}, err
-		}
-		return model.QValue{Kind: model.QValueKindBoolean, Value: val}, nil
-	case pgtype.TextOID, pgtype.VarcharOID:
-		return model.QValue{Kind: model.QValueKindString, Value: string(*rawBytes)}, nil
-	case pgtype.TimestampOID, pgtype.TimestamptzOID:
-		val, err := time.Parse(time.RFC3339, string(*rawBytes))
-		if err != nil {
-			return model.QValue{}, err
-		}
-		et, err := model.NewExtendedTime(val, model.DateTimeKindType, "")
-		if err != nil {
-			return model.QValue{}, fmt.Errorf("failed to create extended time: %w", err)
-		}
-		return model.QValue{Kind: model.QValueKindETime, Value: et}, nil
-	case pgtype.NumericOID:
-		return model.QValue{Kind: model.QValueKindNumeric, Value: string(*rawBytes)}, nil
-	case pgtype.UUIDOID:
-		return model.QValue{Kind: model.QValueKindString, Value: string(*rawBytes)}, nil
-	case pgtype.ByteaOID:
-		return model.QValue{Kind: model.QValueKindBytes, Value: []byte(*rawBytes)}, nil
-	default:
-		typ, _ := pgtype.NewMap().TypeForOID(oid)
-		fmt.Printf("QValueKindInvalid => oid: %v, typename: %v\n", oid, typ)
-		return model.QValue{Kind: model.QValueKindInvalid, Value: nil}, nil
-	}
-}
-
-func (c *PostgresConnector) PullQRepRecords(config *protos.QRepConfig,
-	partition *protos.QRepPartition,
-) (*model.QRecordBatch, error) {
+func (c *PostgresConnector) PullQRepRecords(
+	config *protos.QRepConfig,
+	partition *protos.QRepPartition) (*model.QRecordBatch, error) {
 	var rangeStart interface{}
 	var rangeEnd interface{}
 
@@ -217,8 +110,10 @@ func (c *PostgresConnector) PullQRepRecords(config *protos.QRepConfig,
 	// Be sure to order the results by the watermark column to ensure consistency across pulls
 	query := fmt.Sprintf("%s WHERE %s BETWEEN $1 AND $2", config.Query, config.WatermarkColumn)
 
+	executor := NewQRepQueryExecutor(c.pool, c.ctx)
+
 	// Execute the query with the range values
-	rows, err := c.pool.Query(c.ctx, query, rangeStart, rangeEnd)
+	rows, err := executor.ExecuteQuery(query, rangeStart, rangeEnd)
 	if err != nil {
 		return nil, err
 	}
@@ -230,25 +125,16 @@ func (c *PostgresConnector) PullQRepRecords(config *protos.QRepConfig,
 	// log the number of field descriptions and column names
 	fmt.Printf("field descriptions: %v\n", fieldDescriptions)
 
-	// Initialize the record batch
+	// Process the rows and retrieve the records
+	records, err := executor.ProcessRows(rows, fieldDescriptions)
+	if err != nil {
+		return nil, err
+	}
+
+	// Create the record batch
 	batch := &model.QRecordBatch{
-		NumRecords: 0,
-		Records:    make([]*model.QRecord, 0),
-	}
-
-	// Map each row to a QRecord and add it to the batch
-	for rows.Next() {
-		record, err := mapRowToQRecord(rows, fieldDescriptions)
-		if err != nil {
-			return nil, fmt.Errorf("failed to map row to QRecord: %w", err)
-		}
-		batch.Records = append(batch.Records, record)
-		batch.NumRecords++
-	}
-
-	// Check for any errors encountered during iteration
-	if rows.Err() != nil {
-		return nil, fmt.Errorf("row iteration failed: %w", rows.Err())
+		NumRecords: uint32(len(records)),
+		Records:    records,
 	}
 
 	return batch, nil
